feat(components): Restructure layout components and rename files for consistency
- Moved layout components (e.g., Navbar, Footer) into dedicated folders
- Renamed files to lowercase with "-" as word separator for better readability
- Updated "cv" to "curriculum-vitae" for clarity and consistency
