+++ app/cv/page.tsx
@@ -1,7 +1,7 @@
-import CV from '@/components/layout/CV';
+import CurriculumVitae from '@/components/layout/curriculum-vitae/curriculum-vitae';
 import { Button } from '@/components/ui/button';
 
-export default function CVPage() {
+export default function CurriculumVitaePage() {
   return (
     <div className="w-full max-w-screen-xl mx-auto py-12 px-4 space-y-6">
       <h1 className="text-3xl font-bold text-center">
@@ -14,7 +14,7 @@ export default function CVPage() {
           </a>
         </Button>
       </div>
-      <CV />
+      <CurriculumVitae />
     </div>
   );
 }
