+++ components/layout/projects/project-card.tsx
@@ -7,7 +7,7 @@ import {
   CardTitle,
 } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
-import ProjectImage from './ProjectImage';
+import ProjectImage from '@/components/layout/projects/project-image';
 import { Badge } from '@/components/ui/badge';
 
 interface IProjectCardProps {
