+++ components/layout/projects/project-image.tsx
